Create AnoryTemplateComponent for default app layout; Moves business logic of HomePage to useHomePageViewModel
