+++ frontend/.eslintrc.json
@@ -25,6 +25,7 @@
     "tailwindcss"
   ],
   "rules": {
-    "react/require-default-props": 0
+    "react/require-default-props": 0,
+    "react/prop-types": 0
   }
 }
