+++ frontend/src/Features/Home/useHomePageViewModel.tsx
@@ -1,18 +1,15 @@
-import { useDisclosure, useToast } from '@chakra-ui/react';
+import { useToast } from '@chakra-ui/react';
 import { useEffect, useMemo, useState } from 'react';
 import { AxiosResponse } from 'axios';
 import StoryTileMapper from '../../Mapper/StoryTileMapper';
 import APICallBuilder from '../../UseCases/APICallBuilder';
-import GetCategoriesUseCase from '../../UseCases/Category/GetCategoriesUseCase';
 import CreateStoryUseCase from '../../UseCases/Story/CreateStoryUseCase';
 import GetStoriesUseCase from '../../UseCases/Story/GetStoriesUseCase';
 import { CreateStoryPayload } from '../../UseCases/Story/Payload/CreateStoryPayload';
 
-export default function useHomePageViewModel() {
-  const { isOpen, onOpen, onClose } = useDisclosure();
+export default function useHomePageViewModel(closeModal: () => void) {
   const toast = useToast();
   const [storyData, setStoryData] = useState<any[]>([null, null, null, null, null]);
-  const [availableCategories, setAvailableCategories] = useState<any[]>([]);
 
   const showToast = (status: any, title: String, description: String | null = null) => {
     toast({
@@ -24,14 +21,6 @@ export default function useHomePageViewModel() {
     });
   };
 
-  const onGetCategoriesSuccess = (response: AxiosResponse) => setAvailableCategories(response.data);
-  const onGetCategoriesFailed = (error: any) => showToast('error', error.response.data.message, 'Try refresh this page.');
-
-  const getCategoriesAPI = new APICallBuilder()
-    .addAction(GetCategoriesUseCase.create())
-    .addOnSuccess(onGetCategoriesSuccess)
-    .addOnFailed(onGetCategoriesFailed);
-
   const onGetStoriesSuccess = (response: AxiosResponse) => setStoryData(response.data);
   const onGetStoriesFailed = (error: any) => showToast('error', error.response.data);
 
@@ -42,7 +31,7 @@ export default function useHomePageViewModel() {
     .addOnFailed(onGetStoriesFailed);
 
   const onSubmitStorySuccess = () => {
-    onClose();
+    closeModal();
     showToast('success', 'Story Created');
     getStoriesAPI.call();
   };
@@ -68,23 +57,14 @@ export default function useHomePageViewModel() {
       .call();
   };
 
-  const categoriesElement = useMemo(() => availableCategories.map(
-    (element) => <option key={element.id} value={element.id}>{element.name}</option>,
-  ), [availableCategories]);
-
   const storiesElement = useMemo(() => StoryTileMapper.handle(storyData), [storyData]);
 
   useEffect(() => {
-    getCategoriesAPI.call();
     getStoriesAPI.call();
   }, []);
 
   return {
-    isModalOpen: isOpen,
-    openModal: onOpen,
-    closeModal: onClose,
-    onSubmitStoryClick,
-    categoriesElement,
     storiesElement,
+    onSubmitStoryClick,
   };
 }
